Improve catelogs and schedule selection
