+++ src/components/Carousel/BukkaCard.scss
@@ -22,7 +22,7 @@ $fontFamily: 'Roboto', sans-serif;
 }
 
 .bukka-img-color {
-  background: rgb(233, 231, 231);
+  background-color: rgb(233, 231, 231);
 }
 
 .card-wrap {
@@ -95,7 +95,7 @@ $fontFamily: 'Roboto', sans-serif;
   height: 100%;
   -webkit-transition: opacity 0.4s linear;
   transition: opacity 0.4s linear;
-  // background-color: rgb(233, 231, 231);
+  background-color: rgb(233, 231, 231);
 }
