+++ src/components/common-navs/LocationNavSmallScreen.jsx
@@ -1,85 +1,19 @@
-import React, { Fragment, useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { useMediaQuery } from 'react-responsive';
 import { connect } from 'react-redux';
-import Clock from 'Icons/Clock';
-import ChevronVertical from '../icons/ChevronVertical';
-import DismissModal from '../modal/DismissModal';
-import Modal from '../modal';
+
 import Button from '../button/Button';
-import SearchLocation from '../places-suggest/SearchLocation';
-import DeliveryOrPickupNav from './DeliveryOrPickupNav';
 import ChevronRight from '../icons/ChevronRight';
 import MapMarker from '../icons/MapMarker';
-import { ConnectedDurationContent } from './Duration';
 
 import './LocationNavSmallScreen.scss';
 
-import Cancel from '../icons/Cancel';
 import { useLocationContext } from '../../context/LocationContext';
+import { useModalContext } from '../../context/ModalContext';
 
-const CancelScheduleBtn = ({ handleClick }) => (
-  <div
-    className="dismiss-modal"
-    tabIndex="0"
-    role="button"
-    aria-pressed="false"
-    onClick={handleClick}
-  >
-    <Cancel />
-  </div>
-);
 
-export const SelectLocationModal = ({ delivery, currentSchedule, mode }) => {
-  const [schedule, setSchedule] = useState(false);
-  return (
-    <Modal
-      classNames="select-delivery-pickup"
-      dataTarget="small-location-nav"
-      largeOnSmallScreen
-    >
-      <div className="small-search-container">
-        {!schedule && <DismissModal />}
-        {schedule &&
-        <CancelScheduleBtn handleClick={() => setSchedule(false)} />}
-        <div className="small-search-wrapper">
-          <div className="dropdown-suggestion">
-            {!schedule &&
-            <Fragment>
-              <DeliveryOrPickupNav delivery={delivery} />
-              <SuggestionsDropdown handleClick={() => {}} />
-              <div className="carousel-divider mb-0" />
-              <Button type="button" classNames="schedule-later-btn" handleClick={() => setSchedule(true)}>
-                <span className="col-10 text-left">
-                  <span className="top--1">
-                    <Clock />
-                  </span>
-                  <span className="pl-3 align-bottom">Schedule for Later</span>
-                </span>
-                <span className="col-2">
-                  <ChevronVertical />
-                </span>
-              </Button>
-            </Fragment>
-            }
-            {schedule && <ConnectedDurationContent mode={mode} currentSchedule={currentSchedule} />}
-          </div>
-        </div>
-      </div>
-    </Modal>
-  );
-};
-
-const SuggestionsDropdown = () => (
-  <div className="suggestion-dropdown">
-    <SearchLocation
-      chevronButtonVisible={false}
-      showDeliveryOrPickupNav={false}
-      showDropdown
-    />
-  </div>
-);
-
-const ButtonText = ({ mode }) => {
+const CurrentLocation = ({ handleClick, mode, bukka }) => {
   const { selectedLocation } = useLocationContext();
   const [state, setState] = useState('Current Location');
 
@@ -92,54 +26,61 @@ const ButtonText = ({ mode }) => {
   }, [selectedLocation]);
 
   return (
-    <h2 className="inline-text">
-      {mode === 'delivery' && <span>Delivery to</span>}
-      {mode === 'pickup' && <span>PICKUP NEAR</span>}
-      <span className="text">
-        {state}
-        <span className="chevron-down">
-          <ChevronRight />
-        </span>
-      </span>
-    </h2>
+    <Button
+      type="button"
+      classNames="small-nav-btn container pl-10"
+      handleClick={handleClick}
+    >
+      {bukka ?
+        <h2 className="inline-text">
+          {mode === 'delivery' && <span>Delivery to</span>}
+          {mode === 'pickup' && <span>PICKUP NEAR</span>}
+          <span className="text">
+            {state}
+            <span className="chevron-down">
+              <ChevronRight />
+            </span>
+          </span>
+        </h2>
+        : <h2 className="inline-text">
+          <span className="bukka-inline-text">
+            <span className="chevron-down pr-1">
+              <MapMarker />
+            </span>
+            {state}
+          </span>
+        </h2>}
+    </Button>
   );
 };
 
-const BukkaButtonText = () => (
-  <h2 className="inline-text">
-    <span className="bukka-inline-text">
-      <span className="chevron-down pr-1">
-        <MapMarker />
-      </span>
-    </span>
-  </h2>
-);
-
-const CurrentLocation = ({ handleClick, mode, bukka }) => (
-  <Button
-    type="button"
-    classNames="small-nav-btn container pl-10"
-    handleClick={handleClick}
-  >
-    {!bukka && <ButtonText mode={mode} />}
-    {bukka && <BukkaButtonText />}
-  </Button>
-);
-
 const LocationNavSmallScreen = ({ mode, bukka }) => {
   const { selectedLocation } = useLocationContext();
+  const isMobileScreen = useMediaQuery({ maxWidth: 576 });
+  const { setModal, setSelectLocationPopup } = useModalContext();
+
+  const onClick = () => {
+    setSelectLocationPopup(true);
+    setModal(true);
+  };
+
   return (
-    <Fragment>
+    isMobileScreen &&
       <div
-        className="d-block d-sm-block d-md-none
-      d-lg-none d-xl-none gutter-bg-clor"
+        className="gutter-bg-clor"
       >
         <div className="location-navbar-content container">
           <div className="options-center col-lg-12 px-0">
             <div className="options-wrapper">
               <div className="options">
                 <div className="btn-location">
-                  <div className="small-location-nav" data-target="#small-location-nav" data-toggle="modal">
+                  <div
+                    aria-pressed="true"
+                    tabIndex="0"
+                    role="button"
+                    className="small-location-nav"
+                    onClick={onClick}
+                  >
                     <CurrentLocation
                       mode={mode}
                       bukka={bukka}
@@ -153,7 +94,6 @@ const LocationNavSmallScreen = ({ mode, bukka }) => {
           </div>
         </div>
       </div>
-    </Fragment>
   );
 };
 
@@ -174,24 +114,8 @@ LocationNavSmallScreen.propTypes = {
   bukka: PropTypes.bool
 };
 
-ButtonText.propTypes = {
-  mode: PropTypes.string.isRequired
-};
-
 CurrentLocation.propTypes = {
   handleClick: PropTypes.func.isRequired,
   mode: PropTypes.string.isRequired,
   bukka: PropTypes.bool.isRequired
-};
-
-SelectLocationModal.defaultProps = {
-  delivery: false
-};
-
-SelectLocationModal.propTypes = {
-  delivery: PropTypes.bool
-};
-
-CancelScheduleBtn.propTypes = {
-  handleClick: PropTypes.func.isRequired
 };
