+++ src/components/dropdown/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import PropTypes from 'prop-types';
 
-import './dropdown.scss';
+import './index.scss';
 
 const DropDown = ({ children }) => (
   <div className="dropdown-menu" aria-labelledby="dropdownMenuButton">
