+++ src/components/schedule/index.scss
