+++ src/features/feed/common/areastoexplore.scss
@@ -1,6 +1,6 @@
 .feed-header {
-  position: fixed;
-  top: 65px;
+  // position: fixed;
+  // top: 65px;
   width: 100%;
   background-color: #9c1c26;
   height: 248px;
@@ -12,6 +12,12 @@
 .feed-header-content {
   height: 100%;
   position: relative;
+}
+
+.Feed-Banner-Content {
+  position: absolute;
+  width: 100%;
+  height: 100%;
 }
 
 .header-img-section {
